調整專案整體目錄規劃
1. 新增 - 多項功能用 API、Interfaces 與 Local Files。
2. 調整 - 多項功能用 API、Interfaces 與 Local Files。
3. 調整 - 幾項 class 、函數命名至更貼切其用途。
