+++ src/app/dashboard/dashboard.component.ts
@@ -1,10 +1,19 @@
+
 import { Component, HostListener, OnInit } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { CommonService } from '../shared/common.service';
 import * as _ from 'lodash';
 import { Router } from '@angular/router';
-																									
+
+import { CommonService } from '../shared/common.service';														
 import { LanguageService } from '../shared/service/language.service';
+
+// For import interfaces
+import { FieldSummary }     from '../shared/interfaces/Field/For_queryFieldSummaryInfo';  // @2024/03/14 Add by yuchen
+import { FieldList, Field } from '../shared/interfaces/Field/For_queryFieldList';         // @2024/03/14 Add by yuchen
+
+// For import local files
+import { localFieldSummaryInfo } from '../shared/local-files/Field/For_queryFieldSummaryInfo'; // @2024/03/14 Add by yuchen
+import { localFieldList }        from '../shared/local-files/Field/For_queryFieldList';        // @2024/03/14 Add by yuchen
 
 export interface SystemSummary {
   ocloudCount: number;
@@ -19,213 +28,138 @@ export interface SystemSummary {
   avgNetwork: string;
 }
 
-export interface FieldSummary {
-  fieldSummaryInfo: FieldSummaryInfo[];
-}
-
-export interface FieldSummaryInfo {
-  bsNum: number;
-  fieldNum: number;
-  ueNum: string;
-}
-
-export interface FieldList {
-  fields: Fields[];
-}
-
-export interface Fields {
-  id: string;
-  name: string;
-  phone: string;
-  fieldposition1: string;
-  fieldposition2: string;
-  fieldposition3: string;
-  fieldposition4: string;
-  bsinfo: Bsinfo[];
-  bsNum: number;
-  ueNum: string;
-  coverage: string;
-  accessibility: string;
-  availability: string;
-  mobility: string;
-  retainability: string;
-  energy: string;
-  integrity: Integrity;
-  utilization: Utilization;
-  alarmCriticalNum: number;
-  alarmMajorNum: number;
-  alarmMinorNum: number;
-  alarmWarningNum: number;
-}
-
-export interface Bsinfo {
-  id: string;
-  name: string;
-}
-
-export interface Integrity {
-  downlinkDelay: string;
-  uplinkDelay: string;
-  downlinkThrouthput: string;
-  uplinkThrouthput: string;
-}
-
-export interface Utilization {
-  pdu: string;
-  resourceProcess: string;
-  resourceMemory: string;
-  resourceDisk: string;
-  maxPdu: string;
-}
-
-export interface OcloudSummary {
-  id: string;
-  name: string;
-  dmsCount: number,
-  nfCount: number,
-  faultCount: number,
-  ocloudSummary?: OcloudSummary
-}
-
 @Component({
   selector: 'app-dashboard',
   templateUrl: './dashboard.component.html',
   styleUrls: ['./dashboard.component.scss']
 })
+
 export class DashboardComponent implements OnInit {
+
   sessionId: string = '';
-  systemSummary: SystemSummary = {} as SystemSummary;
-  fieldSummary: FieldSummary = {} as FieldSummary;
-  fieldList: FieldList = {} as FieldList;
-  ueNumList: string[] = [];
-  ocloudSummary: OcloudSummary[] = [];
-  // utilizationPercent: number = 0;
-  resizeTime: any;
-  circularHeight!: number;
-  ueNum: string = '0';
-					  
-  showVircularGauge: boolean = false;
 
   /* CRITICAL,MAJOR,MINOR,WARNING */
   severitys: string[];
 
-  constructor(private http: HttpClient, public commonService: CommonService, private router: Router, public languageService: LanguageService) {
+  constructor(
+
+    private                   http: HttpClient, 
+    public           commonService: CommonService, 
+    private                 router: Router, 
+    public         languageService: LanguageService,
+
+    public fieldSummary_LocalFiles: localFieldSummaryInfo, // fieldSummary_LocalFiles 用於從 Local 文件獲取總結所有場域資訊
+    public    fieldList_LocalFiles: localFieldList,        // fieldList_LocalFiles 用於從 Local 文件獲取場域列表數據
+    
+  ) {
+
     this.severitys = this.commonService.severitys;
   }
 
   ngOnInit(): void {
+
     this.sessionId = this.commonService.getSessionId();
+
     // Field Summary
     this.getfieldSummaryInfo();
     this.getfieldListInfo();
-    // Ocloud Summary
-    //this.getOcloudSummary();
   }
-
-  getSystemSummary() {
-    if (this.commonService.isLocal) {
-      /* local file test */
-      this.systemSummary = this.commonService.systemSummary;
-      this.systemSummaryDeal();
-    } else {
-      const url = `${this.commonService.restPath}/querySystemSummary/${this.sessionId}`;
-      this.http.get(url).subscribe(
-        res => {
-          console.log('getSystemSummary:');
-          console.log(res);
-          this.systemSummary = res as SystemSummary;
-          this.systemSummaryDeal();
-        }
-      );
-    }
-  }
-
+  
+  ueNum: string = '0';
+  fieldSummary: FieldSummary = {} as FieldSummary;
   getfieldSummaryInfo() {
-    if (this.commonService.isLocal) {
-      /* local file test */
-      this.fieldSummary = this.commonService.fieldSummary;
-	    this.fieldSummary.fieldSummaryInfo.forEach(field => {
-        this.ueNum = (field.ueNum.match(/\d+/)?.[0] || '0').toString();
-      //field.ueNum = extractedNumber;
+
+    if ( this.commonService.isLocal ) {
+      // local file test
+
+      this.fieldSummary = this.fieldSummary_LocalFiles.fieldSummary_local;
+	    this.fieldSummary.fieldSummaryInfo.forEach( field => {
+
+        this.ueNum = ( field.ueNum.match(/\d+/)?.[0] || '0' ).toString();
+        //field.ueNum = extractedNumber;
+
       });
-      console.log('fieldSummaryInfo:', this.fieldSummary);				
-      this.systemSummaryDeal();
+
+      console.log( 'fieldSummaryInfo:', this.fieldSummary );		
+
     } else {
+
       const url = `${this.commonService.restPath}/queryFieldSummaryInfo/${this.sessionId}`;
-      this.http.get(url).subscribe(
+
+      this.http.get( url ).subscribe(
         res => {
-          console.log('queryFieldSummaryInfo:');
-          console.log(res);
+          console.log( 'queryFieldSummaryInfo:', res );
           this.fieldSummary = res as FieldSummary;
-          this.systemSummaryDeal();
         }
       );
     }
   }
 
+  fieldList: FieldList = {} as FieldList;
   getfieldListInfo() {
-    if (this.commonService.isLocal) {
+
+    if ( this.commonService.isLocal ) {
       /* local file test */
-      this.fieldList = this.commonService.fieldList;
-      this.systemSummaryDeal();
+
+      this.fieldList = this.fieldList_LocalFiles.fieldList_local;
     } else {
+
       const url = `${this.commonService.restPath}/queryFieldList/${this.sessionId}`;
-      this.http.get(url).subscribe(
+
+      this.http.get( url ).subscribe(
         res => {
-          console.log('queryFieldList:');
-          console.log(res);
+          console.log( 'queryFieldList:', res );
           this.fieldList = res as FieldList;
-          this.systemSummaryDeal();
         }
       );
     }
   }
 
-  NumUes(inputString: string): string {
+  ueNumList: string[] = [];
+  numUes( inputString: string ): string {
+
     const matches = inputString.match(/\d+/);
+
     return matches ? matches[0] : '0';
   }
 
-  systemSummaryDeal() {
-    // this.utilizationPercent = Math.floor((Number(this.systemSummary.usedNodes) / Number(this.systemSummary.totalNodes)) * 100);
-    // window.setTimeout(() => {
-    //   this.showVircularGauge = true;
-    //   this.resize();
-    // }, 0);
-    // if (this.systemSummary.avgCpu != 'N/A' || this.systemSummary.avgStorage != 'N/A' ||
-    //   this.systemSummary.totalMemory != 'N/A' || this.systemSummary.avgNetwork != 'N/A' ) {
-    //     this.systemSummary.avgCpu += ' %';
-    //     this.systemSummary.totalMemory += ' GB';
-    //     this.systemSummary.avgStorage += ' MB';
-    //     this.systemSummary.avgNetwork += ' Kbps';
-    //   }
-  }
-
   // @12/13 Add by yuchen
-  selectField!: Fields;   
+  selectField!: Field;  
+
   // @12/13 Update by yuchen
-  view(fields: Fields) {
+  view( fields: Field ) {
     this.selectField = fields;
-    console.log("View Detail of the field id:", this.selectField.id, "and the field name: ", this.selectField.name);
-    this.router.navigate(['/main/field-mgr/info', this.selectField.id, this.selectField.name]);
+    console.log( "View Detail of the field id:", this.selectField.id, "and the field name: ", this.selectField.name );
+    this.router.navigate( ['/main/field-mgr/info', this.selectField.id, this.selectField.name] );
   }
 
-  severityCount(severity: string): number {
-    if (severity.toUpperCase() === this.severitys[0]) {
+
+  systemSummary: SystemSummary = {} as SystemSummary;
+
+  severityCount( severity: string ): number {
+
+    if ( severity.toUpperCase() === this.severitys[0] ) {
       return this.systemSummary.totalCritical;
-    } else if (severity.toUpperCase() === this.severitys[1]) {
+    } else if ( severity.toUpperCase() === this.severitys[1] ) {
+
       return this.systemSummary.totalMajor;
-    } else if (severity.toUpperCase() === this.severitys[2]) {
+
+    } else if ( severity.toUpperCase() === this.severitys[2] ) {
+
       return this.systemSummary.totalMinor;
-    } else if (severity.toUpperCase() === this.severitys[3]) {
+
+    } else if ( severity.toUpperCase() === this.severitys[3] ) {
+      
       return this.systemSummary.totalWarning;
+
     } else {
+
       return 0;
     }
   }
 
-  severityText(severity: string): string {
-    return this.commonService.severityText(severity);
+  severityText( severity: string ): string {
+    return this.commonService.severityText( severity );
   }
 
 }
