+++ src/app/shared/api/For_Field.ts
@@ -1,26 +1,26 @@
-// 場域頁面會用到的 API 都於此 ( 還未移動完整 ) @2024/03/06 Update
+// Field 頁面會用到的 API 都於此 ( 還未移動完整 ) @2024/03/06 Update
 
 import { HttpHeaders, HttpClient, HttpParams  } from '@angular/common/http';
 import { Injectable }    from '@angular/core';
 import { Observable }    from 'rxjs';
 import { CommonService } from '../common.service'; // @2024/01/08 Add by yuchen
 
-import { BSInfo }      from '../../shared/interfaces/BS/For_queryBsInfo_BS';       // @2024/02/05 Add
-import { BSInfo_dist } from '../../shared/interfaces/BS/For_queryBsInfo_dist_BS';  // @2024/02/05 Add
+import { BSInfo }      from '../interfaces/BS/For_queryBsInfo_BS';       // @2024/02/05 Add
+import { BSInfo_dist } from '../interfaces/BS/For_queryBsInfo_dist_BS';  // @2024/02/05 Add
 
 @Injectable({
   providedIn: 'root'
 })
-export class apiForField {
-
-  restPath = this.commonService.restPath;         // Get the root path  @2024/01/08
-  sessionId = this.commonService.getSessionId();  // Get the Session ID @2024/01/08
+export class apiForFieldMgmt {
 
   constructor(
     private http: HttpClient,
     private commonService: CommonService 
   ) {}
   
+  restPath = this.commonService.restPath;         // Get the root path  @2024/01/08
+  sessionId = this.commonService.getSessionId();  // Get the Session ID @2024/01/08
+
 
   // Get Lists of Field @11/30 Add by yuchen
   queryFieldList(): Observable<any> {
