+++ src/app/shared/common.service.ts
@@ -1,3 +1,4 @@
+
 import { Injectable } from '@angular/core';
 import { HttpHeaders, HttpClient, HttpParams  } from '@angular/common/http';
 import { FormGroup } from '@angular/forms';
@@ -14,23 +15,8 @@ import { DmsAvaliableCapacity, NfCapacityList, NfCapacitySummary, NfInfo, NfPerf
 
 import { Item } from './models/item';
 
-//import { MainComponent } from '../main/main.component';
-
-// Interfaces of Dashboard
-import { SystemSummary } from '../dashboard/dashboard.component';
-import { OcloudSummary } from '../dashboard/dashboard.component';
-import { FieldSummary } from '../dashboard/dashboard.component';
-import { FieldList } from '../dashboard/dashboard.component';
-
-// Interfaces of Field Management
-import { FieldInfo } from './interfaces/Field/For_queryFieldInfo'; // @12/21 Update by yuchen
-
-// Interfaces of BS Management
-import { BSInfo } from './interfaces/BS/For_queryBsInfo_BS';             // @12/21 Update by yuchen
-import { BSInfo_dist } from './interfaces/BS/For_queryBsInfo_dist_BS';   // @12/24 Add
-
 // Interfaces of NE ( Component ) management 
-import { ComponentList } from '../component-management/component-management.component';
+import { ComponentList }   from '../component-management/component-management.component';
 import { BsComponentInfo } from '../component-management/component-info/component-info.component';
 import { ComponentInfosw } from '../component-management/component-info/component-info.component';
 
@@ -54,13 +40,6 @@ import { SoftwareInfo } from '../software-management/software-info/software-info
 import { AccountLists} from '../account-management/account-management.component';
 import { AccountInfo } from '../account-management/account-info/account-info.component';
 
-// Interfaces of Schedule Management
-
-// Interfaces of Log Management
-import { UserLogsList } from '../shared/interfaces/Log/For_queryLogList';        // Update by yutchen @2024/03/13
-import { NELogsList } from '../shared/interfaces/Log/For_queryUserNetconfLog';   // Update by yutchen @2024/03/13
-
-
 export interface NowTime {
   year: string;
   month: string;
@@ -74,17 +53,17 @@ export interface NowTime {
 })
 export class CommonService {
 
-  isLocal!: boolean;
-  restPath!: string;
-  options = { headers: new HttpHeaders({ 'Content-Type': 'multipart/form-data' }) };
-  severitys: string[] = ['CRITICAL', 'MAJOR', 'MINOR', 'WARNING'];
+      isLocal !: boolean;
+     restPath !: string;
+      options = { headers: new HttpHeaders({ 'Content-Type': 'multipart/form-data' }) };
+    severitys: string[] = ['CRITICAL', 'MAJOR', 'MINOR', 'WARNING'];
   statusTypes: string[] = ['New', 'Cleared'];
-  situations: string[] = ['Pending', 'Ended'];
-  cmpsource: string[] = ['Startup', 'Running', 'Candidate'];
+   situations: string[] = ['Pending', 'Ended'];
+    cmpsource: string[] = ['Startup', 'Running', 'Candidate'];
 
   // For Log Management @11/01 Add 
   UserLogType: string[] = ['GET', 'POST', 'DELETE'];
-  NELogType: string[] = ['get', 'get-config', 'edit-config'];
+    NELogType: string[] = ['get', 'get-config', 'edit-config'];
 
   // Software Management
   nfTypeList: Item[] = [
@@ -114,14 +93,20 @@ export class CommonService {
   statusMapDisplayName: Map<string, string> = new Map();
   typeMapDisplayName: Map<string, string> = new Map();
 
-  constructor(private http: HttpClient) {
-    this.statusList.forEach((row) => {
+  constructor ( 
+    
+    private http: HttpClient
+    
+  ) {
+
+    this.statusList.forEach(( row ) => {
       this.statusMapDisplayName.set(row.value, row.displayName);
     });
 
-    this.TypeList.forEach((row) => {
+    this.TypeList.forEach(( row ) => {
       this.typeMapDisplayName.set(row.value, row.displayName);
     });
+
   }
   
   // 調整為 RxJS 新版本( Observer object )語法 @11/28 changed by yuchen
@@ -267,7 +252,8 @@ export class CommonService {
     }
   }
 
-  /* API */
+
+/* API */
 //  loginpage(body: {}): Observable<any> {
 //    const url = `${this.restPath}/ `;
 //    const bodyStr = JSON.stringify(body);
@@ -490,7 +476,7 @@ export class CommonService {
     return this.http.get(url);
   }
 
-  //Component Mgm API
+  // Component Mgmt API
   queryBsComponentList(): Observable<any> {
     const url = `${this.restPath}/queryBsComponentList/${this.getSessionId()}`;
     return this.http.get(url);
@@ -505,465 +491,7 @@ export class CommonService {
     return this.http.get(url);
   }
 
-
-  // Field Management API  @11/30 Add by yuchen
-  queryFieldList(): Observable<any> {
-    
-    // 構建 API URL
-    const url = `${this.restPath}/queryFieldList/${this.getSessionId()}`;
-  
-    // 發起 HTTP GET 請求
-    return this.http.get(url);
-  }
-
-  queryFieldInfo(fieldId: string): Observable<any> {  // @12/05 Add by yuchen
-    
-    // 構建 API URL
-    const url = `${this.restPath}/queryFieldInfo/${this.getSessionId()}/${fieldId}`;
-  
-    // 發起 HTTP GET 請求
-    return this.http.get(url);
-  }
-
-  // get the Field Image @2024/01/02 Add by yuchen
-  queryFieldImage(fieldId: string): Observable<any> { 
-    
-    // 構建 API URL
-    const url = `${this.restPath}/queryFieldImage/${this.getSessionId()}/${fieldId}`;
-  
-    // 發起 HTTP GET 請求
-    return this.http.get(url);
-  }
-
-  // Get SINR or RSRP map @2024/01/04 Update by yuchen
-  bsHeatMap( fieldId: string, leftLongitude: number, leftLatitude: number, rightLongitude: number,
-             rightLatitude: number, mapType: number ): Observable<any> {
-    
-    // 構建 API URL，指向後端的 bsHeatMap 路徑
-    const url = `${this.restPath}/bsHeatMap`;
-  
-    // 準備請求體，包含所有後端所需的參數
-    const requestBody = {
-      session: this.getSessionId(),
-      fieldId: fieldId,
-      'left-longitude': String(Math.round(leftLongitude * 1000000)),    // 轉為字符串
-      'left-latitude': String(Math.round(rightLatitude * 1000000)),     // 轉為字符串
-      'right-longitude': String(Math.round(rightLongitude * 1000000)),  // 轉為字符串
-      'right-latitude': String(Math.round(leftLatitude * 1000000)),     // 轉為字符串
-      maptype: mapType
-    };
-    
-    console.log( "session:", requestBody.session, 
-                  "\nfieldId:", requestBody.fieldId, 
-                    "\nmaptype:", requestBody.maptype, 
-                    '\nleft-latitude:', requestBody['left-latitude'],
-                      '\nleft-longitude:', requestBody['left-longitude'], 
-                      '\nright-latitude:', requestBody['right-latitude'], 
-                        '\nright-longitude:', requestBody['right-longitude'] );
-
-    console.log( "JSON to bsHeatMap:", requestBody );
-
-    // 發起 HTTP POST 請求，並返回 Observable 以便訂閱和處理響應
-    return this.http.post( url, requestBody );
-  }
-
-
-  // Schedule  Management API  @11/24 Add by yuchen
-  createJobTicket(body: any): Observable<any> {        // 建立 Schedule 用
-    const url = `${this.restPath}/createJobTicket/${this.getSessionId()}`;
-    const bodyStr = JSON.stringify(body);
-    return this.http.post(url, bodyStr);
-  }
-  removeJobTicket(jobId: string): Observable<any> {    // 移除現有 Schedule 用
-    const url = `${this.restPath}/removeJobTicket/${this.getSessionId()}/${jobId}`;
-    return this.http.delete(url);
-  }
-  queryJobTicketInfo(jobId: string): Observable<any> { // 取得現有 Schedule 資訊用
-    const url = `${this.restPath}/queryJobTicketInfo/${this.getSessionId()}/${jobId}`;
-    return this.http.get(url);
-  }
-  getReportFile(jobticketId: string): Observable<any> { // 下載 Schedule 報表用
-    const url = `${this.restPath}/getReportFile/${this.getSessionId()}/${jobticketId}`;
-    return this.http.get(url);
-  }
-  queryJobTicketList(): Observable<any> {               // 取得現有的 Schedule List 用
-    const url = `${this.restPath}/queryJobTicketList/${this.getSessionId()}`;
-    return this.http.get(url);
-  }
-
-  // Log Management API @2024/03/12 Update by yuchen
-
-  // 取得 User Logs 用
-  queryLogList( params: any ): Observable<any> {
-    const url = `${this.restPath}/queryLogList/${this.getSessionId()}`;
-    return this.http.get( url, { params } );  // 將參數物件傳遞給 HTTP GET 請求
-  }
-
-  // 下載 User Logs 用 @2024/03/12 Add
-  getDumpLogList( params: any ): Observable<any> {
-    const url = `${this.restPath}/getDumpLogList/${this.getSessionId()}`;
-    return this.http.get( url, { params } );  // 將參數物件傳遞給 HTTP GET 請求
-  }
-
-  // 取得 NE Logs 用
-  queryUserNetconfLog( params: any ): Observable<any> { 
-    const url = `${this.restPath}/queryUserNetconfLog/${this.getSessionId()}`;
-    return this.http.get( url, { params } );  // 將參數物件傳遞給 HTTP GET 請求
-  }
-
-  // 下載 NE Logs 用 @2024/03/12 Add
-  getDumpUserNetconfLog( params: any ): Observable<any> {
-    const url = `${this.restPath}/getDumpUserNetconfLog/${this.getSessionId()}`;
-    return this.http.get( url, { params } );  // 將參數物件傳遞給 HTTP GET 請求
-  }
-
-
-  /* local file */
-  systemSummary: SystemSummary = {
-    ocloudCount: 2,
-    nfCount: 10,
-    totalCritical: 20,
-    totalMajor: 10,
-    totalMinor: 15,
-    totalWarning: 30,
-    avgCpu: "15%",
-    totalMemory: "72 GB",
-    avgStorage: "70.123 KBps",
-    avgNetwork: "3.475 KBps"
-  };
-
-  fieldSummary: FieldSummary = {
-    fieldSummaryInfo: [
-      {
-        "bsNum": 1,
-        "fieldNum": 2,
-        "ueNum": "itri 13"
-      }
-    ]
-  };
-
-  // @12/07 Update by yuchen
-  fieldInfo: FieldInfo = {
-
-        "id": "aa520ec8e5074f54b77d",
-        "name": "itri-field1",
-        "phone": "0912345678",
-        "fieldposition1": "[121.049825,24.778037]",
-        "fieldposition2": "[121.039552,24.778037]",
-        "fieldposition3": "[121.039552,24.770575]",
-        "fieldposition4": "[121.049825,24.770575]",
-        "bsinfo": [
-          {
-            "id": "97a606c4994344e09d3b",
-            "name": "BS01",
-            "accessibility": "100.0",
-            "mobility": "74.37",
-            "retainability": "0.0",
-            "energy": "28780.54",
-            "integrity": {
-              "downlinkDelay": "0",
-              "uplinkDelay": "0",
-              "downlinkThrouthput": "0",
-              "uplinkThrouthput": "0"
-            },
-            "utilization": {
-              "pdu": "0",
-              "resourceProcess": "5.0",
-              "resourceMemory": "47.0",
-              "resourceDisk": "0.0",
-              "maxPdu": "623"
-            }
-          },
-          {
-            "id": "c9bb081540ea4752982f",
-            "name": "BS02",
-            "accessibility": null,
-            "mobility": null,
-            "retainability": null,
-            "energy": null,
-            "integrity": {
-              "downlinkDelay": null,
-              "uplinkDelay": null,
-              "downlinkThrouthput": null,
-              "uplinkThrouthput": null
-            },
-            "utilization": {
-              "pdu": null,
-              "resourceProcess": "20",
-              "resourceMemory": "11",
-              "resourceDisk": null,
-              "maxPdu": null
-            },
-            "cellInfo": [
-              {
-                "nci": "000050108",
-                "accessibility": "100.0",
-                "mobility": "75.15",
-                "retainability": "0.0",
-                "energy": "28780.54",
-                "integrity": {
-                  "downlinkDelay": "0",
-                  "uplinkDelay": "0",
-                  "downlinkThrouthput": "0",
-                  "uplinkThrouthput": "0"
-                },
-                "utilization": {
-                  "pdu": "0",
-                  "resourceProcess": "5.0",
-                  "resourceMemory": "47.0",
-                  "resourceDisk": "0.0",
-                  "maxPdu": "624"
-                }
-              },
-              {
-                "nci": "000050109",
-                "accessibility": "100.0",
-                "mobility": "80.84",
-                "retainability": "0.0",
-                "energy": "28780.54",
-                "integrity": {
-                  "downlinkDelay": "0",
-                  "uplinkDelay": "0",
-                  "downlinkThrouthput": "0",
-                  "uplinkThrouthput": "0"
-                },
-                "utilization": {
-                  "pdu": "0",
-                  "resourceProcess": "5.0",
-                  "resourceMemory": "47.0",
-                  "resourceDisk": "0.0",
-                  "maxPdu": "624"
-                }
-              }
-            ]
-          }
-        ],
-        "bsNum": 2,
-        "ueNum": "fake 10",
-        "coverage": "100",
-        "accessibility": "100.0",
-        "availability": "99.98",
-        "mobility": "76.73",
-        "retainability": "0.0",
-        "energy": "28780.54",
-        "integrity": {
-          "downlinkDelay": "0",
-          "uplinkDelay": "0",
-          "downlinkThrouthput": "0",
-          "uplinkThrouthput": "0"
-        },
-        "utilization": {
-          "pdu": "0",
-          "resourceProcess": "5.0",
-          "resourceMemory": "47.0",
-          "resourceDisk": "0.0",
-          "maxPdu": "1871"
-        },
-        "alarmCriticalNum": 5,
-        "alarmMajorNum": 294448,
-        "alarmMinorNum": 740791,
-        "alarmWarningNum": 0
-  }
-
-
-  fieldList: FieldList = {
-    fields: [
-      {
-        "id": "aa520ec8e5074f54b77d",
-        "name": "itri-field1",
-        "phone": "0912345678",
-        "fieldposition1": "[121.044734,24.774227]",
-        "fieldposition2": "[121.043587,24.774227]",
-        "fieldposition3": "[121.043587,24.773396]",
-        "fieldposition4": "[121.044734,24.773396]",
-        bsinfo: [
-          {
-            "id": "fd005824997144b987d5",
-            "name": "bs_1x2x4"
-          }
-        ],
-        "bsNum": 1,
-        "ueNum": "fake 0",
-        "coverage": "15",
-        "accessibility": "DRB Accessibility:14",
-        "availability": "string",
-        "mobility": "NG-RAN handover success rate:78.21",
-        "retainability": "DRB Retainability:0.36",
-        "energy": "NG-RAN data Energy Efficiency:27.67",
-        integrity: {
-          "downlinkDelay": "6504264",
-          "uplinkDelay": "6507722",
-          "downlinkThrouthput": "1617437",
-          "uplinkThrouthput": "1629079"
-        },
-        utilization: {
-          "pdu": "1627179",
-          "resourceProcess": "29.1",
-          "resourceMemory": "60.43",
-          "resourceDisk": "3.05",
-          "maxPdu": "1674863"
-        },
-        "alarmCriticalNum": 36106,
-        "alarmMajorNum": 289582,
-        "alarmMinorNum": 557217,
-        "alarmWarningNum": 18212
-      },
-      {
-        "id": "aa520ec8e5074f54b77d",
-        "name": "itri-field2",
-        "phone": "0912345678",
-        "fieldposition1": "[121.044734,24.774227]",
-        "fieldposition2": "[121.043587,24.774227]",
-        "fieldposition3": "[121.043587,24.773396]",
-        "fieldposition4": "[121.044734,24.773396]",
-        bsinfo: [
-          {
-            "id": "fd005824997144b987d5",
-            "name": "bs_1x2x4"
-          }
-        ],
-        "bsNum": 2,
-        "ueNum": "itri 2",
-        "coverage": "10",
-        "accessibility": "DRB Accessibility:23",
-        "availability": "string",
-        "mobility": "NG-RAN handover success rate:81.36",
-        "retainability": "DRB Retainability:0.33",
-        "energy": "NG-RAN data Energy Efficiency:25.35",
-        integrity: {
-          "downlinkDelay": "6502345",
-          "uplinkDelay": "6506657",
-          "downlinkThrouthput": "1617241",
-          "uplinkThrouthput": "1626573"
-        },
-        utilization: {
-          "pdu": "1626024",
-          "resourceProcess": "31.2",
-          "resourceMemory": "72.15",
-          "resourceDisk": "5.31",
-          "maxPdu": "1674863"
-        },
-        "alarmCriticalNum": 100,
-        "alarmMajorNum": 200,
-        "alarmMinorNum": 300,
-        "alarmWarningNum": 100
-      },
-      {
-        "id": "aa520ec8e5074f54b77d",
-        "name": "itri-field3",
-        "phone": "0912345678",
-        "fieldposition1": "[121.044734,24.774227]",
-        "fieldposition2": "[121.043587,24.774227]",
-        "fieldposition3": "[121.043587,24.773396]",
-        "fieldposition4": "[121.044734,24.773396]",
-        bsinfo: [
-          {
-            "id": "fd005824997144b987d5",
-            "name": "bs_1x2x4"
-          }
-        ],
-        "bsNum": 2,
-        "ueNum": "itri 3",
-        "coverage": "10",
-        "accessibility": "DRB Accessibility:51",
-        "availability": "string",
-        "mobility": "NG-RAN handover success rate:77.38",
-        "retainability": "DRB Retainability:0.32",
-        "energy": "NG-RAN data Energy Efficiency:29.56",
-        integrity: {
-          "downlinkDelay": "6487754",
-          "uplinkDelay": "6495574",
-          "downlinkThrouthput": "1606875",
-          "uplinkThrouthput": "1618045"
-        },
-        utilization: {
-          "pdu": "1605781",
-          "resourceProcess": "25.3",
-          "resourceMemory": "52.78",
-          "resourceDisk": "2.85",
-          "maxPdu": "1678754"
-        },
-        "alarmCriticalNum": 100,
-        "alarmMajorNum": 200,
-        "alarmMinorNum": 300,
-        "alarmWarningNum": 100
-      }
-    ]
-  };
-
-  ocloudSummary: OcloudSummary[] = [
-    {
-      id: "cloud000-0000-0000-0000-000000000000",
-      name: "cloud0",
-      dmsCount: 1,
-      nfCount: 2,
-      faultCount: 20
-    },
-    {
-      id: "cloud000-0000-0000-0000-000000000001",
-      name: "cloud1",
-      dmsCount: 1,
-      nfCount: 0,
-      faultCount: 0
-    },
-    {
-      id: "cloud000-0000-0000-0000-000000000002",
-      name: "cloud2",
-      dmsCount: 2,
-      nfCount: 0,
-      faultCount: 0
-    },
-    {
-      id: "cloud000-0000-0000-0000-000000000003",
-      name: "cloud3",
-      dmsCount: 1,
-      nfCount: 0,
-      faultCount: 0
-    },
-    {
-      id: "cloud000-0000-0000-0000-000000000004",
-      name: "cloud4",
-      dmsCount: 2,
-      nfCount: 0,
-      faultCount: 0
-    },
-    {
-      id: "cloud000-0000-0000-0000-000000000005",
-      name: "cloud5",
-      dmsCount: 1,
-      nfCount: 0,
-      faultCount: 0
-    },
-    {
-      id: "cloud000-0000-0000-0000-000000000006",
-      name: "cloud6",
-      dmsCount: 2,
-      nfCount: 0,
-      faultCount: 0
-    }
-    ,
-    {
-      id: "cloud000-0000-0000-0000-000000000007",
-      name: "cloud7",
-      dmsCount: 1,
-      nfCount: 0,
-      faultCount: 0
-    },
-    {
-      id: "cloud000-0000-0000-0000-000000000008",
-      name: "cloud8",
-      dmsCount: 2,
-      nfCount: 0,
-      faultCount: 0
-    },
-    {
-      id: "cloud000-0000-0000-0000-000000000009",
-      name: "cloud9",
-      dmsCount: 2,
-      nfCount: 0,
-      faultCount: 0
-    }
-  ];
+// Local Files ↓
 
   ocloudList: OCloudList[] = [
     {
@@ -4614,329 +4142,4 @@ export class CommonService {
     acknowledgeOwner: "Sam"
   }
 
-  // User logs files @12/21 Update by yutchen add 5 files
-  UserLogsList: UserLogsList = {
-  
-    logNumber: 20,  // number
-    loginfo: [
-      {
-        userid: "k200",   // string
-        logtype: "POST",  // string
-        loglevel: 20,     // number
-        logmsg: "k200 login success session: irm_session_0309079f",  // string
-        logtime: "2024-02-14 13:59:46"  // string
-      },
-      {
-        userid: "k300",
-        logtype: "GET",
-        loglevel: 20,
-        logmsg: "queryUploadFileList success field Id: k300",
-        logtime: "2024-02-27 17:44:19"
-      },
-      {
-        userid: "k200",
-        logtype: "GET",
-        loglevel: 20,
-        logmsg: "k200 queryJobTicketList success",
-        logtime: "2024-02-29 17:21:04"
-      },
-      {
-        userid: "k100",
-        logtype: "GET",
-        loglevel: 20,
-        logmsg: "QueryBsList success field Id: k200",
-        logtime: "2024-02-16 16:48:34"
-      },
-      {
-        userid: "k200",
-        logtype: "GET",
-        loglevel: 20,
-        logmsg: "k200 QueryBsComponentList Success",
-        logtime: "2024-02-19 16:48:34"
-      },
-      {
-        userid: "k200",
-        logtype: "GET",
-        loglevel: 20,
-        logmsg: "k200 queryFieldList success",
-        logtime: "2024-03-06 16:48:34"
-      },
-      {
-        userid: "k000",
-        logtype: "GET",
-        loglevel: 20,
-        logmsg: "k000 queryFieldList success",
-        logtime: "2024-03-01 16:40:37"
-      },
-      {
-        userid: "k200",
-        logtype: "GET",
-        loglevel: 20,
-        logmsg: "queryUploadFileList success field Id: k200",
-        logtime: "2024-03-03 16:48:38"
-      },
-      {
-        userid: "k200",
-        logtype: "GET",
-        loglevel: 20,
-        logmsg: "k200 queryJobTicketList success",
-        logtime: "2024-02-21 14:48:39"
-      },
-      {
-        userid: "k100",
-        logtype: "GET",
-        loglevel: 20,
-        logmsg: "k100 QueryBsComponentList Success",
-        logtime: "2024-02-23 16:48:47"
-      },
-      {
-        userid: "k200",
-        logtype: "DELETE",
-        loglevel: 20,
-        logmsg: "RemoveField success field Id: 1a933410af39457eb2c8",
-        logtime: "2024-02-23 16:35:52"
-      },
-      {
-        userid: "k100",
-        logtype: "GET",
-        loglevel: 20,
-        logmsg: "k100 queryFieldList success",
-        logtime: "2024-03-03 16:35:53"
-      },
-      {
-        userid: "k000",
-        logtype: "GET",
-        loglevel: 20,
-        logmsg: "QueryBsList success field Id: k000",
-        logtime: "2024-02-29 16:35:54"
-      },
-      {
-        userid: "k300",
-        logtype: "DELETE",
-        loglevel: 20,
-        logmsg: "RemoveBs success Id: k300",
-        logtime: "2024-02-29 16:35:59"
-      },
-      {
-        userid: "k100",
-        logtype: "POST",
-        loglevel: 20,
-        logmsg: "queryUploadFileList success field Id: k100",
-        logtime: "2024-03-04 16:40:19"
-      },      
-      {
-        userid: "k000",
-        logtype: "GET",
-        loglevel: 20,
-        logmsg: "queryUploadFileList success field Id: k000",
-        logtime: "2024-03-04 17:2:19"
-      },
-      {
-        userid: "k100",
-        logtype: "GET",
-        loglevel: 20,
-        logmsg: "k100 queryJobTicketList success",
-        logtime: "2024-03-04 18:19:04"
-      },      
-      {
-        userid: "k000",
-        logtype: "GET",
-        loglevel: 20,
-        logmsg: "k000 queryFieldList success",
-        logtime: "2024-03-04 13:48:37"
-      },
-      {
-        userid: "k200",
-        logtype: "GET",
-        loglevel: 20,
-        logmsg: "queryUploadFileList success field Id: k200",
-        logtime: "2024-02-20 13:48:38"
-      },
-      {
-        userid: "k200",
-        logtype: "GET",
-        loglevel: 20,
-        logmsg: "k200 queryJobTicketList success",
-        logtime: "2024-02-19 21:48:39"
-      },
-    ]
-  };
-
-  // NE logs files @12/21 Update by yutchen add 5 files
-  NELogsList: NELogsList = {
-    
-    logNumber: 21,
-    loginfo: [
-      {
-        userid: "k200",             // string
-        operation: "get-config",    // string
-        // string
-        req_data: "<filter><software-inventory xmlns=\"urn:itri-software-management\"><software-slot></software-slot></software-inventory></filter>",
-        // string
-        resp_data: "<?xml version=\"1.0\" ?>\n<data xmlns=\"urn:ietf:params:xml:ns:netconf:base:1.0\">\n\t<software-inventory xmlns=\"urn:itri-software-management\">\n\t\t<software-slot>\n\t\t\t<name>slot-1</name>\n\t\t\t<status>VALID</status>\n\t\t\t<active>true</active>\n\t\t\t<running>true</running>\n\t\t\t<access>READ_ONLY</access>\n\t\t\t<vendor-code>K2</vendor-code>\n\t\t\t<build-id>b01</build-id>\n\t\t\t<build-name>product-default</build-name>\n\t\t\t<build-version>0.1.0</build-version>\n\t\t\t<files>\n\t\t\t\t<name>file-1</name>\n\t\t\t\t<version>0.2.3</version>\n\t\t\t\t<local-path>~/some_dir/</local-path>\n\t\t\t\t<integrity>OK</integrity>\n\t\t\t</files>\n\t\t</software-slot>\n\t\t<software-slot>\n\t\t\t<name>slot-2</name>\n\t\t\t<status>EMPTY</status>\n\t\t\t<active>false</active>\n\t\t\t<running>false</running>\n\t\t\t<access>READ_WRITE</access>\n\t\t</software-slot>\n\t\t<software-slot>\n\t\t\t<name>slot-3</name>\n\t\t\t<status>EMPTY</status>\n\t\t\t<active>false</active>\n\t\t\t<running>false</running>\n\t\t\t<access>READ_WRITE</access>\n\t\t</software-slot>\n\t\t<software-slot>\n\t\t\t<name>slot-4</name>\n\t\t\t<status>EMPTY</status>\n\t\t\t<active>false</active>\n\t\t\t<running>false</running>\n\t\t\t<access>READ_WRITE</access>\n\t\t</software-slot>\n\t</software-inventory>\n</data>\n",
-        logtime: "2024-03-04 13:59:46",   // string
-        comp_name: "itri_10.0.2.17"       // string @11/22 Add 
-      },
-      {              
-        userid: "k300",
-        operation: "get-config",
-        req_data: "<filter><software-inventory xmlns=\"urn:itri-software-management\"><software-slot></software-slot></software-inventory></filter>",
-        resp_data: "<?xml version=\"1.0\" ?>\n<data xmlns=\"urn:ietf:params:xml:ns:netconf:base:1.0\">",
-        logtime: "2024-03-04 17:44:19",   // string
-        comp_name: "itri_10.0.2.17"       // string @11/22 Add 
-      },
-      { 
-        userid: "k200",
-        operation: "get-config",
-        req_data: "",
-        resp_data: "<?xml version=\"1.0\" ?>\n<data xmlns=\"urn:ietf:params:xml:ns:netconf:base:1.0\">\n\t<software-inventory xmlns=\"urn:o-ran:software-management:1.0\">\n\t\t<software-slot>\n\t\t\t<name>slot-1</name>\n\t\t\t<status>VALID</status>\n\t\t\t<active>true</active>\n\t\t\t<running>true</running>\n\t\t\t<access>READ_ONLY</access>\n\t\t\t<vendor-code>K2</vendor-code>\n\t\t\t<build-id>b01</build-id>\n\t\t\t<build-name>product-default</build-name>\n\t\t\t<build-version>0.1.0</build-version>\n\t\t\t<files>\n\t\t\t\t<name>file-1</name>\n\t\t\t\t<version>0.2.3</version>\n\t\t\t\t<local-path>~/some_dir/</local-path>\n\t\t\t\t<integrity>OK</integrity>\n\t\t\t</files>\n\t\t</software-slot>\n\t\t<software-slot>\n\t\t\t<name>slot-2</name>\n\t\t\t<status>EMPTY</status>\n\t\t\t<active>false</active>\n\t\t\t<running>false</running>\n\t\t\t<access>READ_WRITE</access>\n\t\t</software-slot>\n\t\t<software-slot>\n\t\t\t<name>slot-3</name>\n\t\t\t<status>EMPTY</status>\n\t\t\t<active>false</active>\n\t\t\t<running>false</running>\n\t\t\t<access>READ_WRITE</access>\n\t\t</software-slot>\n\t\t<software-slot>\n\t\t\t<name>slot-4</name>\n\t\t\t<status>EMPTY</status>\n\t\t\t<active>false</active>\n\t\t\t<running>false</running>\n\t\t\t<access>READ_WRITE</access>\n\t\t</software-slot>\n\t</software-inventory>\n</data>\n",
-        logtime: "2024-03-04 17:21:04",         // string
-        comp_name: "itri_10.0.2.17"  // string @11/22 Add 
-      },
-      {
-        userid: "k100",
-        operation: "edit-config",
-        req_data: "<config> <SMO xmlns=\"urn:rdns:com:itri:nr-smo:1.0\"> <ftpip>10.0.2.16</ftpip> <ftpid>k200</ftpid> <ftpkey>k200123</ftpkey> <ftppath>/home/k200/irm-itri/server/uploadFolder/ftpFolder/46a315326d5b4ac2a3a7</ftppath> <file-upload-interval>60</file-upload-interval> </SMO> </config>",
-        resp_data: "1",
-        logtime: "2024-02-29 17:10:23.979271",   // string
-        comp_name: "itri_10.0.2.17"   // string @11/22 Add 
-      },
-      {
-        userid: "k200",
-        operation: "edit-config",
-        req_data: "<config> <SMO xmlns=\"urn:rdns:com:itri:nr-smo:1.0\"> <ftpip>10.0.2.16</ftpip> <ftpid>k200</ftpid> <ftpkey>k200123</ftpkey> <ftppath>/home/k200/irm-itri/server/uploadFolder/ftpFolder/635778a42e854c17a4c2</ftppath> <file-upload-interval>60</file-upload-interval> </SMO> </config>",
-        resp_data: "1",
-        logtime: "2024-02-29 17:10:22.793374",  // string
-        comp_name: "itri_10.0.2.16"  // string @11/22 Add 
-      },
-      {
-        userid: "k200",
-        operation: "edit-config",
-        req_data: "<config> <SMO xmlns=\"urn:rdns:com:itri:nr-smo:1.0\"> <ftpip>10.0.2.16</ftpip> <ftpid>k200</ftpid> <ftpkey>k200123</ftpkey> <ftppath>/home/k200/irm-itri/server/uploadFolder/ftpFolder/614de0d30d1845129f02</ftppath> <file-upload-interval>60</file-upload-interval> </SMO> </config>",
-        resp_data: "1",
-        logtime: "2024-02-29 17:10:21.286927",  // string
-        comp_name: "itri_10.0.2.16"  // string @11/22 Add 
-      },
-      {
-        userid: "k000",
-        operation: "edit-config",
-        req_data: "<config><ME xmlns=\"urn:3gpp:tsg:sa5:nrm:ngran\"><GNBCUFunction><id>0</id><NRCellCU><id>1</id><vsDataContainer><id>0</id><vsData><![CDATA[ <gnbvs xmlns=\"urn:rdns:com:radisys:nr:gnb\"> <gnbCuConfig> <id>0</id> <gnbCellCuVsConfig> <id>0</id> <duId>1</duId> <nRpCI>1</nRpCI> <ueInactivityTimerSec>day30</ueInactivityTimerSec> <eutraConfig> <eutraNeighbourCell> <eutraCellIdentifier>2</eutraCellIdentifier> <MCC>466</MCC> <MNC>66</MNC> <enbIdType>MACRO_ENB_ID</enbIdType> <enbId>1</enbId> </eutraNeighbourCell> <eutraQoSConfig> <configIndex>1</configIndex> <qci>1</qci> <snSizeDL>12</snSizeDL> <snSizeUL>12</snSizeUL> <ulDataSplitThresholdInBytes>b100</ulDataSplitThresholdInBytes> <enableUlOutOfOrderDelivery>false</enableUlOutOfOrderDelivery> <rlcMode>RLC_UM</rlcMode> <rlcUmDir>RLC_UM_DIR_BIDIRECTIONAL</rlcUmDir> <reorderingTimerMs>500</reorderingTimerMs> </eutraQoSConfig> </eutraConfig> <nrConfig> <NRNeighbourCell> <nrCellIdentifier>000004108</nrCellIdentifier> <nrPci>100</nrPci> <MCC>466</MCC> <MNC>55</MNC> <nrModeTdd> <TDDInfo> <nrArfcn>0</nrArfcn> </TDDInfo> </nrModeTdd> <nrArfcn>4850</nrArfcn> <TAC>111111</TAC> </NRNeighbourCell> <NRNeighbourCell> <nrCellIdentifier>000050108</nrCellIdentifier> <nrPci>160</nrPci> <MCC>466</MCC> <MNC>55</MNC> <nrModeTdd> <TDDInfo> <nrArfcn>0</nrArfcn> </TDDInfo> </nrModeTdd> <nrArfcn>4850</nrArfcn> <TAC>111111</TAC> </NRNeighbourCell> </nrConfig> <ueCapabilityTriggerAfterSMCProc>true</ueCapabilityTriggerAfterSMCProc> </gnbCellCuVsConfig> </gnbCuConfig> </gnbvs> ]]></vsData></vsDataContainer></NRCellCU></GNBCUFunction></ME></config>",
-        resp_data: "1",
-        logtime: "2024-02-29 17:08:08.866697",  // string
-        comp_name: "itri_10.0.2.18"  // string @11/22 Add 
-      },
-      {
-        userid: "k200",
-        operation: "get-config",
-        req_data: "<filter><software-inventory xmlns=\"urn:itri-software-management\"><software-slot></software-slot></software-inventory></filter>",
-        resp_data: "<?xml version=\"1.0\" ?> <data xmlns=\"urn:ietf:params:xml:ns:netconf:base:1.0\"> <software-inventory xmlns=\"urn:itri-software-management\"> <software-slot> <name>slot-1</name> <status>VALID</status> <active>false</active> <running>true</running> <access>READ_ONLY</access> <vendor-code>K2</vendor-code> <build-id>b01</build-id> <build-name>product-default</build-name> <build-version>0.1.0</build-version> <files> <name>file-1</name> <version>0.2.3</version> <local-path>~/some_dir/</local-path> <integrity>OK</integrity> </files> </software-slot> <software-slot> <name>slot-2</name> <status>VALID</status> <active>true</active> <running>false</running> <access>READ_WRITE</access> <vendor-code>WN</vendor-code> <build-id>1</build-id> <build-name>WNC-official-build</build-name> <build-version>1-0-0</build-version> <files> <name>fw-v1-0-0.bin</name> <version>1.0.0</version> <local-path>/sw_inventory/slot_3/</local-path> <integrity>OK</integrity> </files> </software-slot> <software-slot> <name>slot-3</name> <status>EMPTY</status> <active>false</active> <running>false</running> <access>READ_WRITE</access> </software-slot> <software-slot> <name>slot-4</name> <status>EMPTY</status> <active>false</active> <running>false</running> <access>READ_WRITE</access> </software-slot> </software-inventory> </data>",
-        logtime: "2024-03-01 17:15:31.669459",  // string
-        comp_name: "itri_10.0.2.19"  // string @11/22 Add 
-      },
-      {
-        userid: "k200",
-        operation: "get-config",
-        req_data: "<filter><software-inventory xmlns=\"urn:itri-software-management\"><software-slot></software-slot></software-inventory></filter>",
-        resp_data: "<?xml version=\"1.0\" ?> <data xmlns=\"urn:ietf:params:xml:ns:netconf:base:1.0\"> <software-inventory xmlns=\"urn:itri-software-management\"> <software-slot> <name>slot-1</name> <status>VALID</status> <active>true</active> <running>true</running> <access>READ_ONLY</access> <vendor-code>K2</vendor-code> <build-id>b01</build-id> <build-name>product-default</build-name> <build-version>0.1.0</build-version> <files> <name>file-1</name> <version>0.2.3</version> <local-path>~/some_dir/</local-path> <integrity>OK</integrity> </files> </software-slot> <software-slot> <name>slot-2</name> <status>EMPTY</status> <active>false</active> <running>false</running> <access>READ_WRITE</access> </software-slot> <software-slot> <name>slot-3</name> <status>EMPTY</status> <active>false</active> <running>false</running> <access>READ_WRITE</access> </software-slot> <software-slot> <name>slot-4</name> <status>EMPTY</status> <active>false</active> <running>false</running> <access>READ_WRITE</access> </software-slot> </software-inventory> </data>",
-        logtime: "2024-03-01 17:15:11.961997",  // string
-        comp_name: "itri_10.0.2.19"  // string @11/22 Add 
-      },
-      {
-        userid: "k100",
-        operation: "get-config",
-        req_data: "<ME xmlns=\"urn:3gpp:tsg:sa5:nrm:ngran\"></ME>",
-        resp_data: "<?xml version=\"1.0\" ?> <data xmlns=\"urn:ietf:params:xml:ns:netconf:base:1.0\"/>",
-        logtime: "2024-03-01 17:10:36.391256",  // string
-        comp_name: "itri_10.0.2.18"  // string @11/22 Add 
-      },
-      {
-        userid: "k200",
-        operation: "get-config",
-        req_data: "<filter><software-inventory xmlns=\"urn:itri-software-management\"><software-slot></software-slot></software-inventory></filter>",
-        resp_data: "<?xml version=\"1.0\" ?> <data xmlns=\"urn:ietf:params:xml:ns:netconf:base:1.0\"> <software-inventory xmlns=\"urn:itri-software-management\"> <software-slot> <name>slot-1</name> <status>VALID</status> <active>true</active> <running>true</running> <access>READ_ONLY</access> <vendor-code>K2</vendor-code> <build-id>b01</build-id> <build-name>product-default</build-name> <build-version>0.1.0</build-version> <files> <name>file-1</name> <version>0.2.3</version> <local-path>~/some_dir/</local-path> <integrity>OK</integrity> </files> </software-slot> <software-slot> <name>slot-2</name> <status>EMPTY</status> <active>false</active> <running>false</running> <access>READ_WRITE</access> </software-slot> <software-slot> <name>slot-3</name> <status>EMPTY</status> <active>false</active> <running>false</running> <access>READ_WRITE</access> </software-slot> <software-slot> <name>slot-4</name> <status>EMPTY</status> <active>false</active> <running>false</running> <access>READ_WRITE</access> </software-slot> </software-inventory> </data>",
-        logtime: "2024-03-01 17:09:41.933556",  // string
-        comp_name: "itri_10.0.2.15"  // string @11/22 Add 
-      },
-      {
-        userid: "k100",
-        operation: "edit-config",
-        req_data: "<config><ME xmlns=\"urn:3gpp:tsg:sa5:nrm:ngran\" ><GNBDUFunction><id>0</id><NRCellDU><id>0</id><bWP>1</bWP></NRCellDU></GNBDUFunction></ME></config>",
-        resp_data: "1",
-        logtime: "2024-03-05 17:03:27.174804",  // string
-        comp_name: "itri_10.0.2.15"  // string @11/22 Add 
-      },
-      {
-        userid: "k000",
-        operation: "edit-config",
-        req_data: "<config><ME xmlns=\"urn:3gpp:tsg:sa5:nrm:ngran\" ><GNBDUFunction><id>0</id><NRCellDU><id>0</id><NRSectorCarrier>0</NRSectorCarrier></NRCellDU></GNBDUFunction></ME></config>",
-        resp_data: "1",
-        logtime: "2024-03-05 17:03:25.265801",  // string
-        comp_name: "itri_10.0.2.19"  // string @11/22 Add 
-      },
-      {
-        userid: "k300",
-        operation: "edit-config",
-        req_data: "<config><ME xmlns=\"urn:3gpp:tsg:sa5:nrm:ngran\" ><GNBCUFunction><id>0</id><NRCellCU><id>0</id><s-NSSAI>0</s-NSSAI></NRCellCU></GNBCUFunction></ME></config>",
-        resp_data: "1",
-        logtime: "2024-03-05 17:02:15.773772",  // string
-        comp_name: "itri_10.0.2.19"  // string @11/22 Add 
-      },
-      {
-        userid: "k100",
-        operation: "get-config",
-        req_data: "",
-        resp_data: "<active-alarm-list xmlns=\"urn:o-ran:fm:1.0\"> <active-alarms> <fault-id>1</fault-id> <fault-source>fault-source_1</fault-source> <affected-objects> <name>affected-objects_1-1</name> </affected-objects> <affected-objects> <name>affected-objects_1-2</name> </affected-objects> <fault-severity>MAJOR</fault-severity> <is-cleared>false</is-cleared> <fault-text>fault-text_1</fault-text> <event-time>2020-01-01T00:00:00Z</event-time> </active-alarms> <active-alarms> <fault-id>2</fault-id> <fault-source>fault-source_2</fault-source> <affected-objects> <name>affected-objects_2-1</name> </affected-objects> <affected-objects> <name>affected-objects_2-2</name> </affected-objects> <fault-severity>MINOR</fault-severity> <is-cleared>false</is-cleared> <fault-text>fault-text_2</fault-text> <event-time>2020-01-11T00:00:00Z</event-time> </active-alarms> <active-alarms> <fault-id>3</fault-id> <fault-source>fault-source_3</fault-source> <affected-objects> <name>affected-objects_3-1</name> </affected-objects> <affected-objects> <name>affected-objects_3-2</name> </affected-objects> <fault-severity>CRITICAL</fault-severity> <is-cleared>false</is-cleared> <fault-text>fault-text_3</fault-text> <event-time>2020-01-01T00:00:00Z</event-time> </active-alarms> </active-alarm-list>",
-        logtime: "2024-03-06 17:03:04.549773",  // string
-        comp_name: "itri_10.0.2.20"  // string @11/22 Add 
-      },     
-      {
-        userid: "k200",
-        operation: "get-config",
-        req_data: "<filter><software-inventory xmlns=\"urn:itri-software-management\"><software-slot></software-slot></software-inventory></filter>",
-        resp_data: "<?xml version=\"1.0\" ?> <data xmlns=\"urn:ietf:params:xml:ns:netconf:base:1.0\"> <software-inventory xmlns=\"urn:itri-software-management\"> <software-slot> <name>slot-1</name> <status>VALID</status> <active>true</active> <running>true</running> <access>READ_ONLY</access> <vendor-code>K2</vendor-code> <build-id>b01</build-id> <build-name>product-default</build-name> <build-version>0.1.0</build-version> <files> <name>file-1</name> <version>0.2.3</version> <local-path>~/some_dir/</local-path> <integrity>OK</integrity> </files> </software-slot> <software-slot> <name>slot-2</name> <status>EMPTY</status> <active>false</active> <running>false</running> <access>READ_WRITE</access> </software-slot> <software-slot> <name>slot-3</name> <status>EMPTY</status> <active>false</active> <running>false</running> <access>READ_WRITE</access> </software-slot> <software-slot> <name>slot-4</name> <status>EMPTY</status> <active>false</active> <running>false</running> <access>READ_WRITE</access> </software-slot> </software-inventory> </data>",
-        logtime: "2024-03-06 17:09:41.933556",  // string
-        comp_name: "itri_10.0.2.15"  // string @11/22 Add 
-      },
-      {
-        userid: "k100",
-        operation: "edit-config",
-        req_data: "<config><ME xmlns=\"urn:3gpp:tsg:sa5:nrm:ngran\" ><GNBDUFunction><id>0</id><NRCellDU><id>0</id><bWP>1</bWP></NRCellDU></GNBDUFunction></ME></config>",
-        resp_data: "1",
-        logtime: "2024-02-18 17:00:27.174804",  // string
-        comp_name: "itri_10.0.2.10"  // string @11/22 Add 
-      },
-      {
-        userid: "k000",
-        operation: "edit-config",
-        req_data: "<config><ME xmlns=\"urn:3gpp:tsg:sa5:nrm:ngran\" ><GNBDUFunction><id>0</id><NRCellDU><id>0</id><NRSectorCarrier>0</NRSectorCarrier></NRCellDU></GNBDUFunction></ME></config>",
-        resp_data: "1",
-        logtime: "2024-03-02 16:03:25.265801",  // string
-        comp_name: "itri_10.0.2.12"  // string @11/22 Add 
-      },      
-      {
-        userid: "k000",
-        operation: "edit-config",
-        req_data: "<config><ME xmlns=\"urn:3gpp:tsg:sa5:nrm:ngran\"><GNBCUFunction><id>0</id><NRCellCU><id>1</id><vsDataContainer><id>0</id><vsData><![CDATA[ <gnbvs xmlns=\"urn:rdns:com:radisys:nr:gnb\"> <gnbCuConfig> <id>0</id> <gnbCellCuVsConfig> <id>0</id> <duId>1</duId> <nRpCI>1</nRpCI> <ueInactivityTimerSec>day30</ueInactivityTimerSec> <eutraConfig> <eutraNeighbourCell> <eutraCellIdentifier>2</eutraCellIdentifier> <MCC>466</MCC> <MNC>66</MNC> <enbIdType>MACRO_ENB_ID</enbIdType> <enbId>1</enbId> </eutraNeighbourCell> <eutraQoSConfig> <configIndex>1</configIndex> <qci>1</qci> <snSizeDL>12</snSizeDL> <snSizeUL>12</snSizeUL> <ulDataSplitThresholdInBytes>b100</ulDataSplitThresholdInBytes> <enableUlOutOfOrderDelivery>false</enableUlOutOfOrderDelivery> <rlcMode>RLC_UM</rlcMode> <rlcUmDir>RLC_UM_DIR_BIDIRECTIONAL</rlcUmDir> <reorderingTimerMs>500</reorderingTimerMs> </eutraQoSConfig> </eutraConfig> <nrConfig> <NRNeighbourCell> <nrCellIdentifier>000004108</nrCellIdentifier> <nrPci>100</nrPci> <MCC>466</MCC> <MNC>55</MNC> <nrModeTdd> <TDDInfo> <nrArfcn>0</nrArfcn> </TDDInfo> </nrModeTdd> <nrArfcn>4850</nrArfcn> <TAC>111111</TAC> </NRNeighbourCell> <NRNeighbourCell> <nrCellIdentifier>000050108</nrCellIdentifier> <nrPci>160</nrPci> <MCC>466</MCC> <MNC>55</MNC> <nrModeTdd> <TDDInfo> <nrArfcn>0</nrArfcn> </TDDInfo> </nrModeTdd> <nrArfcn>4850</nrArfcn> <TAC>111111</TAC> </NRNeighbourCell> </nrConfig> <ueCapabilityTriggerAfterSMCProc>true</ueCapabilityTriggerAfterSMCProc> </gnbCellCuVsConfig> </gnbCuConfig> </gnbvs> ]]></vsData></vsDataContainer></NRCellCU></GNBCUFunction></ME></config>",
-        resp_data: "1",
-        logtime: "2024-03-03 17:08:08.866697",  // string
-        comp_name: "itri_10.0.2.18"  // string @11/22 Add 
-      },
-      {
-        userid: "k200",
-        operation: "get-config",
-        req_data: "<filter><software-inventory xmlns=\"urn:itri-software-management\"><software-slot></software-slot></software-inventory></filter>",
-        resp_data: "<?xml version=\"1.0\" ?> <data xmlns=\"urn:ietf:params:xml:ns:netconf:base:1.0\"> <software-inventory xmlns=\"urn:itri-software-management\"> <software-slot> <name>slot-1</name> <status>VALID</status> <active>false</active> <running>true</running> <access>READ_ONLY</access> <vendor-code>K2</vendor-code> <build-id>b01</build-id> <build-name>product-default</build-name> <build-version>0.1.0</build-version> <files> <name>file-1</name> <version>0.2.3</version> <local-path>~/some_dir/</local-path> <integrity>OK</integrity> </files> </software-slot> <software-slot> <name>slot-2</name> <status>VALID</status> <active>true</active> <running>false</running> <access>READ_WRITE</access> <vendor-code>WN</vendor-code> <build-id>1</build-id> <build-name>WNC-official-build</build-name> <build-version>1-0-0</build-version> <files> <name>fw-v1-0-0.bin</name> <version>1.0.0</version> <local-path>/sw_inventory/slot_3/</local-path> <integrity>OK</integrity> </files> </software-slot> <software-slot> <name>slot-3</name> <status>EMPTY</status> <active>false</active> <running>false</running> <access>READ_WRITE</access> </software-slot> <software-slot> <name>slot-4</name> <status>EMPTY</status> <active>false</active> <running>false</running> <access>READ_WRITE</access> </software-slot> </software-inventory> </data>",
-        logtime: "2024-02-20 19:15:31.669459",  // string
-        comp_name: "itri_10.0.2.33"  // string @11/22 Add 
-      },
-      {
-        userid: "k100",
-        operation: "get-config",
-        req_data: "<filter><software-inventory xmlns=\"urn:itri-software-management\"><software-slot></software-slot></software-inventory></filter>",
-        resp_data: "<?xml version=\"1.0\" ?> <data xmlns=\"urn:ietf:params:xml:ns:netconf:base:1.0\"> <software-inventory xmlns=\"urn:itri-software-management\"> <software-slot> <name>slot-1</name> <status>VALID</status> <active>true</active> <running>true</running> <access>READ_ONLY</access> <vendor-code>K2</vendor-code> <build-id>b01</build-id> <build-name>product-default</build-name> <build-version>0.1.0</build-version> <files> <name>file-1</name> <version>0.2.3</version> <local-path>~/some_dir/</local-path> <integrity>OK</integrity> </files> </software-slot> <software-slot> <name>slot-2</name> <status>EMPTY</status> <active>false</active> <running>false</running> <access>READ_WRITE</access> </software-slot> <software-slot> <name>slot-3</name> <status>EMPTY</status> <active>false</active> <running>false</running> <access>READ_WRITE</access> </software-slot> <software-slot> <name>slot-4</name> <status>EMPTY</status> <active>false</active> <running>false</running> <access>READ_WRITE</access> </software-slot> </software-inventory> </data>",
-        logtime: "2024-02-24 17:15:11.961997",  // string
-        comp_name: "itri_10.0.2.12"  // string @11/22 Add 
-      },
-    ]
-  };
 }
