+++ src/app/shared/interfaces/Field/For_queryFieldSummaryInfo.ts
@@ -1,0 +1,13 @@
+
+// @2024/03/14 Add by yuchen
+// 用於儲存總結所有場域資訊
+
+export interface FieldSummary {
+    fieldSummaryInfo: FieldSummaryInfo[];
+}
+  
+export interface FieldSummaryInfo {
+    bsNum: number;
+    fieldNum: number;
+    ueNum: string;
+}
