+++ src/app/shared/local-files/Field/For_queryFieldInfo.ts
@@ -1,0 +1,131 @@
+ 
+ // Interfaces of FieldInfo @2024/03/14 Add
+ import { FieldInfo } from '../../interfaces/Field/For_queryFieldInfo';
+
+ // Local Files for general FieldInfo @2024/03/14 Add 
+ export class localFieldInfo {
+
+    fieldInfo_local: FieldInfo = {
+
+        "id": "aa520ec8e5074f54b77d",
+        "name": "itri-field1",
+        "phone": "0912345678",
+        "fieldposition1": "[121.049825,24.778037]",
+        "fieldposition2": "[121.039552,24.778037]",
+        "fieldposition3": "[121.039552,24.770575]",
+        "fieldposition4": "[121.049825,24.770575]",
+        "bsinfo": [
+          {
+            "id": "97a606c4994344e09d3b",
+            "name": "BS01",
+            "accessibility": "100.0",
+            "mobility": "74.37",
+            "retainability": "0.0",
+            "energy": "28780.54",
+            "integrity": {
+              "downlinkDelay": "0",
+              "uplinkDelay": "0",
+              "downlinkThrouthput": "0",
+              "uplinkThrouthput": "0"
+            },
+            "utilization": {
+              "pdu": "0",
+              "resourceProcess": "5.0",
+              "resourceMemory": "47.0",
+              "resourceDisk": "0.0",
+              "maxPdu": "623"
+            }
+          },
+          {
+            "id": "c9bb081540ea4752982f",
+            "name": "BS02",
+            "accessibility": null,
+            "mobility": null,
+            "retainability": null,
+            "energy": null,
+            "integrity": {
+              "downlinkDelay": null,
+              "uplinkDelay": null,
+              "downlinkThrouthput": null,
+              "uplinkThrouthput": null
+            },
+            "utilization": {
+              "pdu": null,
+              "resourceProcess": "20",
+              "resourceMemory": "11",
+              "resourceDisk": null,
+              "maxPdu": null
+            },
+            "cellInfo": [
+              {
+                "nci": "000050108",
+                "accessibility": "100.0",
+                "mobility": "75.15",
+                "retainability": "0.0",
+                "energy": "28780.54",
+                "integrity": {
+                  "downlinkDelay": "0",
+                  "uplinkDelay": "0",
+                  "downlinkThrouthput": "0",
+                  "uplinkThrouthput": "0"
+                },
+                "utilization": {
+                  "pdu": "0",
+                  "resourceProcess": "5.0",
+                  "resourceMemory": "47.0",
+                  "resourceDisk": "0.0",
+                  "maxPdu": "624"
+                }
+              },
+              {
+                "nci": "000050109",
+                "accessibility": "100.0",
+                "mobility": "80.84",
+                "retainability": "0.0",
+                "energy": "28780.54",
+                "integrity": {
+                  "downlinkDelay": "0",
+                  "uplinkDelay": "0",
+                  "downlinkThrouthput": "0",
+                  "uplinkThrouthput": "0"
+                },
+                "utilization": {
+                  "pdu": "0",
+                  "resourceProcess": "5.0",
+                  "resourceMemory": "47.0",
+                  "resourceDisk": "0.0",
+                  "maxPdu": "624"
+                }
+              }
+            ]
+          }
+        ],
+        "bsNum": 2,
+        "ueNum": "fake 10",
+        "coverage": "100",
+        "accessibility": "100.0",
+        "availability": "99.98",
+        "mobility": "76.73",
+        "retainability": "0.0",
+        "energy": "28780.54",
+        "integrity": {
+          "downlinkDelay": "0",
+          "uplinkDelay": "0",
+          "downlinkThrouthput": "0",
+          "uplinkThrouthput": "0"
+        },
+        "utilization": {
+          "pdu": "0",
+          "resourceProcess": "5.0",
+          "resourceMemory": "47.0",
+          "resourceDisk": "0.0",
+          "maxPdu": "1871"
+        },
+        "alarmCriticalNum": 5,
+        "alarmMajorNum": 294448,
+        "alarmMinorNum": 740791,
+        "alarmWarningNum": 0
+    }
+
+ }
+
