+++ src/app/shared/local-files/Field/For_queryFieldSummaryInfo.ts
@@ -1,0 +1,18 @@
+ 
+ // Interfaces of FieldSummaryInfo @2024/03/14 Add
+ import { FieldSummary } from '../../interfaces/Field/For_queryFieldSummaryInfo';
+
+ // Local Files for general FieldSummaryInfo @2024/03/14 Add 
+ export class localFieldSummaryInfo {
+
+    fieldSummary_local: FieldSummary = {
+        fieldSummaryInfo: [
+          {
+            "bsNum": 1,
+            "fieldNum": 2,
+            "ueNum": "itri 13"
+          }
+        ]
+      };
+
+ }
