+++ src/app/shared/local-files/Log/For_queryLogList.ts
@@ -1,0 +1,190 @@
+
+// Interfaces of Log Management - queryLogList @2024/03/14 Add
+import { UserLogsList } from '../../interfaces/Log/For_queryLogList';
+
+ // Local Files for general UserLogsList @2024/03/14 Add 
+ export class localUserLogsList {
+
+      userLogsList_local: UserLogsList = {
+      
+        logNumber: 25,  // number
+        loginfo: [
+          {
+            userid: "k200",   // string
+            logtype: "POST",  // string
+            loglevel: 20,     // number
+            logmsg: "k200 login success session: irm_session_0309079f",  // string
+            logtime: "2024-02-14 13:59:46"  // string
+          },
+          {
+            userid: "k300",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "queryUploadFileList success field Id: k300",
+            logtime: "2024-02-27 17:44:19"
+          },
+          {
+            userid: "k200",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "k200 queryJobTicketList success",
+            logtime: "2024-02-29 17:21:04"
+          },
+          {
+            userid: "k100",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "QueryBsList success field Id: k200",
+            logtime: "2024-02-16 16:48:34"
+          },
+          {
+            userid: "k200",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "k200 QueryBsComponentList Success",
+            logtime: "2024-02-19 16:48:34"
+          },
+          {
+            userid: "k200",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "k200 queryFieldList success",
+            logtime: "2024-03-06 16:48:34"
+          },
+          {
+            userid: "k000",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "k000 queryFieldList success",
+            logtime: "2024-03-01 16:40:37"
+          },
+          {
+            userid: "k200",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "queryUploadFileList success field Id: k200",
+            logtime: "2024-03-03 16:48:38"
+          },
+          {
+            userid: "k200",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "k200 queryJobTicketList success",
+            logtime: "2024-02-21 14:48:39"
+          },
+          {
+            userid: "k100",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "k100 QueryBsComponentList Success",
+            logtime: "2024-02-23 16:48:47"
+          },
+          {
+            userid: "k200",
+            logtype: "DELETE",
+            loglevel: 20,
+            logmsg: "RemoveField success field Id: 1a933410af39457eb2c8",
+            logtime: "2024-02-23 16:35:52"
+          },
+          {
+            userid: "k100",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "k100 queryFieldList success",
+            logtime: "2024-03-03 16:35:53"
+          },
+          {
+            userid: "k000",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "QueryBsList success field Id: k000",
+            logtime: "2024-02-29 16:35:54"
+          },
+          {
+            userid: "k300",
+            logtype: "DELETE",
+            loglevel: 20,
+            logmsg: "RemoveBs success Id: k300",
+            logtime: "2024-02-29 16:35:59"
+          },
+          {
+            userid: "k100",
+            logtype: "POST",
+            loglevel: 20,
+            logmsg: "queryUploadFileList success field Id: k100",
+            logtime: "2024-03-04 16:40:19"
+          },      
+          {
+            userid: "k000",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "queryUploadFileList success field Id: k000",
+            logtime: "2024-03-04 17:2:19"
+          },
+          {
+            userid: "k100",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "k100 queryJobTicketList success",
+            logtime: "2024-03-04 18:19:04"
+          },      
+          {
+            userid: "k000",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "k000 queryFieldList success",
+            logtime: "2024-03-04 13:48:37"
+          },
+          {
+            userid: "k200",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "queryUploadFileList success field Id: k200",
+            logtime: "2024-02-20 13:48:38"
+          },
+          {
+            userid: "k200",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "k200 queryJobTicketList success",
+            logtime: "2024-02-19 21:48:39"
+          },
+          {
+            userid: "k000",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "queryUploadFileList success field Id: k000",
+            logtime: "2024-03-04 17:2:19"
+          },
+          {
+            userid: "k100",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "k100 queryJobTicketList success",
+            logtime: "2024-03-10 18:19:04"
+          },      
+          {
+            userid: "k000",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "k000 queryFieldList success",
+            logtime: "2024-03-14 13:48:37"
+          },
+          {
+            userid: "k200",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "queryUploadFileList success field Id: k200",
+            logtime: "2024-03-02 13:48:38"
+          },
+          {
+            userid: "k200",
+            logtype: "GET",
+            loglevel: 20,
+            logmsg: "k200 queryJobTicketList success",
+            logtime: "2024-03-05 21:48:39"
+          }
+        ]
+      };
+
+ }
